+++ src/ITesterExtension.php
@@ -10,6 +10,8 @@ namespace MyTester;
  */
 interface ITesterExtension
 {
+    public function onExtensionsLoaded(Events\ExtensionsLoaded $event): void;
+
     public function onTestsStarted(Events\TestsStarted $event): void;
 
     public function onTestsFinished(Events\TestsFinished $event): void;
@@ -17,4 +19,6 @@ interface ITesterExtension
     public function onTestCaseStarted(Events\TestCaseStarted $event): void;
 
     public function onTestCaseFinished(Events\TestCaseFinished $event): void;
+
+    public function getName(): string;
 }